Split Workflow.cpp test file in multiple files
